+++ inst/app/init.R
@@ -5,9 +5,9 @@ r_url_list[["(Dis)similarity"]] <-
 r_url_list[["Attributes"]] <-
   list("tabs_pmap" = list("Summary" = "multivariate/pmap/", "Plot" = "multivariate/pmap/plot/"))
 r_url_list[["Hierarchical"]] <-
-  list("tabs_hier_clus" = list("Summary" = "multivariate/hier_clus/", "Plot" = "multivariate/hier_clus/plot/"))
-r_url_list[["K-means"]] <-
-  list("tabs_kmeans_clus" = list("Summary" = "multivariate/kmeans/", "Plot" = "multivariate/kmeans/plot/"))
+  list("tabs_hclus" = list("Summary" = "multivariate/hclus/", "Plot" = "multivariate/hclus/plot/"))
+r_url_list[["K-clustering"]] <-
+  list("tabs_kclus" = list("Summary" = "multivariate/kclus/", "Plot" = "multivariate/kclus/plot/"))
 r_url_list[["Conjoint"]] <-
   list("tabs_conjoint" = list("Summary" = "multivariate/conjoint/", "Plot" = "multivariate/plot/"))
 options(radiant.url.list = r_url_list); rm(r_url_list)
@@ -23,8 +23,8 @@ options(radiant.multivariate_ui =
       tabPanel("Pre-factor", uiOutput("pre_factor")),
       tabPanel("Factor", uiOutput("full_factor")),
 	    "----", "Cluster",
-    	tabPanel("Hierarchical", uiOutput("hier_clus")),
-    	tabPanel("K-means", uiOutput("kmeans_clus")),
+    	tabPanel("Hierarchical", uiOutput("hclus")),
+    	tabPanel("K-clustering", uiOutput("kclus")),
 	    "----", "Conjoint",
 	    tabPanel("Conjoint", uiOutput("conjoint"))
     )