+++ inst/app/tools/analysis/kclus_ui.R
@@ -1,11 +1,12 @@
 ###############################################################
-# Kmeans clustering
+# K-clustering
 ###############################################################
 
 km_plots <- c("Density" = "density", "Bar" = "bar", "Scatter" = "scatter")
+km_algorithm <- c("K-means" = "mean", "K-medians" = "median")
 
 # list of function arguments
-km_args <- as.list(formals(kmeans_clus))
+km_args <- as.list(formals(kclus))
 
 km_inputs <- reactive({
   # loop needed because reactive values don't allow single bracket indexing
@@ -18,20 +19,31 @@ km_inputs <- reactive({
 
 output$ui_km_vars <- renderUI({
 
- 	isNum <- "numeric" == .getclass() | "integer" == .getclass()
- 	vars <- varnames()[isNum]
+  ## are there any two-level vars
+  dum <- two_level_vars()
+  if (length(dum) > 0) {
+    isVars <- .getclass() %in% c("integer","numeric","factor")
+    isFct <- {.getclass() == "factor"} %>% {names(.[.])} %>% setdiff(., dum)
+    vars <- varnames()[isVars] %>% .[!. %in% isFct]
+  } else {
+    isVars <- .getclass() %in% c("integer","numeric")
+    vars <- varnames()[isVars]
+  }
+
   selectInput(inputId = "km_vars", label = "Variables:", choices = vars,
 	  selected = state_multiple("km_vars", vars, input$hc_vars),
 	  multiple = TRUE, size = min(8, length(vars)), selectize = FALSE)
 })
 
-output$ui_kmeans_clus <- renderUI({
+output$ui_kclus <- renderUI({
   req(input$dataset)
   tagList(
     wellPanel(
       actionButton("km_run", "Estimate", width = "100%")
     ),
   	wellPanel(
+      selectInput("km_fun", label = "Algorithm:", choices = km_algorithm,
+        selected = state_single("km_fun", km_algorithm, "mean"), multiple = FALSE),
 	    uiOutput("ui_km_vars"),
 		  checkboxInput(inputId = "km_hc_init", label = "Initial centers from HC",
       	value = state_init('km_hc_init',FALSE)),
@@ -49,7 +61,7 @@ output$ui_kmeans_clus <- renderUI({
 		  ),
 	    numericInput("km_nr_clus", "Number of clusters:", min = 2,
 	    	value = state_init('km_nr_clus',2)),
-      conditionalPanel(condition = "input.tabs_kmeans_clus == 'Plot'",
+      conditionalPanel(condition = "input.tabs_kclus == 'Plot'",
         selectInput("km_plots", label = "Plot(s):", choices = km_plots,
                  selected = state_multiple("km_plots", km_plots, "density"),
                  multiple = FALSE)
@@ -61,9 +73,9 @@ output$ui_kmeans_clus <- renderUI({
         )
 		  )
   	),
-  	help_and_report(modal_title = "K-means cluster analysis",
-  	                fun_name = "kmeans_clus",
-  	                help_file = inclMD(file.path(getOption("radiant.path.multivariate"),"app/tools/help/kmeans_clus.md")))
+  	help_and_report(modal_title = "K-clustering",
+  	                fun_name = "kclus",
+  	                help_file = inclMD(file.path(getOption("radiant.path.multivariate"),"app/tools/help/kclus.md")))
 	)
 })
 
@@ -78,27 +90,27 @@ km_plot_height <- function()
   km_plot() %>% { if (is.list(.)) .$plot_height else 400 }
 
 # output is called from the main radiant ui.R
-output$kmeans_clus <- renderUI({
+output$kclus <- renderUI({
 
-		register_print_output("summary_kmeans_clus", ".summary_kmeans_clus")
-		register_plot_output("plot_kmeans_clus", ".plot_kmeans_clus",
+		register_print_output("summary_kclus", ".summary_kclus")
+		register_plot_output("plot_kclus", ".plot_kclus",
                          	width_fun = "km_plot_width",
                          	height_fun = "km_plot_height")
 
 	  km_output_panels <- tabsetPanel(
-	    id = "tabs_kmeans_clus",
+	    id = "tabs_kclus",
 	    tabPanel("Summary",
         downloadLink("dl_km_means", "", class = "fa fa-download alignright"), br(),
-        verbatimTextOutput("summary_kmeans_clus")),
+        verbatimTextOutput("summary_kclus")),
 	    tabPanel("Plot",
-        plot_downloader("kmeans_clus", height = km_plot_height()),
-        plotOutput("plot_kmeans_clus", height = "100%"))
+        plot_downloader("kclus", height = km_plot_height()),
+        plotOutput("plot_kclus", height = "100%"))
 	  )
 
 		stat_tab_panel(menu = "Multivariate > Cluster",
-		              tool = "K-means",
-		              tool_ui = "ui_kmeans_clus",
-		             	output_panels = km_output_panels)
+		               tool = "K-clustering",
+		               tool_ui = "ui_kclus",
+		             	 output_panels = km_output_panels)
 })
 
 .km_available <- reactive({
@@ -109,24 +121,24 @@ output$kmeans_clus <- renderUI({
   "available"
 })
 
-.kmeans_clus <- eventReactive(input$km_run, {
+.kclus <- eventReactive(input$km_run, {
   withProgress(message = 'Estimating cluster solution', value = 1,
     # do.call(ann, ann_inputs())
-    do.call(kmeans_clus, km_inputs())
+    do.call(kclus, km_inputs())
   )
 })
 
-.summary_kmeans_clus <- reactive({
+.summary_kclus <- reactive({
   if (.km_available() != "available") return(.km_available())
-  summary(.kmeans_clus())
+  summary(.kclus())
 })
 
-.plot_kmeans_clus <- reactive({
+.plot_kclus <- reactive({
   if (.km_available() != "available") return(.km_available())
-  plot(.kmeans_clus(), plots = input$km_plots, shiny = TRUE)
+  plot(.kclus(), plots = input$km_plots, shiny = TRUE)
 })
 
-observeEvent(input$kmeans_clus_report, {
+observeEvent(input$kclus_report, {
   outputs <- c("summary")
   inp_out <- list(list(dec = 2), "")
   figs <- FALSE
@@ -136,18 +148,18 @@ observeEvent(input$kmeans_clus_report, {
     inp_out[[2]] <- list(plots = input$km_plots)
   }
   update_report(inp_main = clean_args(km_inputs(), km_args),
-                fun_name = "kmeans_clus", inp_out = inp_out,
+                fun_name = "kclus", inp_out = inp_out,
                 outputs = outputs, figs = figs,
                 fig.width = km_plot_width(),
                 fig.height = km_plot_height(),
-                xcmd = paste0("# store(result, name = '", input$km_store_name,"')\n# write.csv(result$clus_means, file = '~/kmeans.csv')"))
+                xcmd = paste0("# store(result, name = '", input$km_store_name,"')\n# write.csv(result$clus_means, file = '~/kclus.csv')"))
 })
 
 output$dl_km_means <- downloadHandler(
-  filename = function() { "kmeans.csv" },
+  filename = function() { "kclus.csv" },
   content = function(file) {
     if (pressed(input$km_run)) {
-      .kmeans_clus() %>% { if (is.list(.)) write.csv(.$clus_means, file = file) }
+      .kclus() %>% { if (is.list(.)) write.csv(.$clus_means, file = file) }
     } else {
       cat("No output available. Press the Estimate button to generate the cluster solution", file = file)
     }
@@ -157,6 +169,6 @@ output$dl_km_means <- downloadHandler(
 ## store cluster membership
 observeEvent(input$km_store, {
   if (pressed(input$km_run)) {
-    .kmeans_clus() %>% { if (is.list(.)) store(., name = input$km_store_name) }
+    .kclus() %>% { if (is.list(.)) store(., name = input$km_store_name) }
   }
 })